Move the login without access step to common steps
